Rename internal modules
